Support 'adb reboot fastboot' command conditionally.
Devices not supporting logical partitions do not require the command.

Bug: 78793464
Test: 'adb reboot fastboot' issued causes a simple reboot for a device
not supporting logical partitions.

+++ init/reboot.cpp
@@ -474,7 +474,8 @@ bool HandlePowerctlMessage(const std::string& command) {
                                << err;
                 }
             } else if (reboot_target == "sideload" || reboot_target == "sideload-auto-reboot" ||
-                       reboot_target == "fastboot") {
+                       (reboot_target == "fastboot" &&
+                        android::base::GetBoolProperty("ro.boot.logical_partitions", false))) {
                 std::string arg = reboot_target == "sideload-auto-reboot" ? "sideload_auto_reboot"
                                                                           : reboot_target;
                 const std::vector<std::string> options = {
